dicom-viewer/proxy/requirements.txt
flask
azure-identity
azure-storage-file-datalake
pydicom
